+++ packages/share_plus/share_plus/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.5.0
+
+- Add `shareXFiles` implementations
+- Deprecate `shareFiles*` implementations
+- Enable `shareXFiles` implementations on Web
+
 ## 4.4.0
 
 - Reverted changes in 4.2.0 due to crash issues. See #1081