+++ packages/share_plus/share_plus_platform_interface/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.1.0
+
+- Add `shareXFiles` implementations
+- Deprecate `shareFiles*` implementations
+- Enable `shareXFiles` implementations on Web
+
 ## 3.0.3
 
 - Update dependencies